Enable item search using simple date strings
As requested in #29, this commit extends the datetime argument
of item searching to accept what I'm calling "simple date strings",
e.g.:

- `2017`
- `2017-06`
- `2017-06-10`

Ranges are also ok:

- `2017/2018`
- `2017-06/2017-07`
- `2017-06-10/2017-06-11`

Simple date strings are expanded to encompass their entire date
resolution. E.g. `2017` expands to the entire year of 2017. The unit
tests and documentation describe and enforce all expected behaviors.

This also includes a unit test to assert failure in the case of three or
more components of a datetime specifier (e.g. `2017/2018/2019` is
illegal).