Allow providing filter-lang to search
The user can optionally provide a filter-lang parameter to the
ItemSearch endpoint when using filter. This preserves the existing
behavior of defaulting to `cql-json` when no value is provided, so it
should be backwards compatible with existing code. By optionally
providing a new value, such as `cql2-json`, the user can supply filters
with different JSON language variants.