Added ItemSearch.item_collection
Adds an ItemSearch.item_collection method as a replacement for
the deprecated `get_all_items()`.

A few other notes:

1. I haven't implemented any of the `max_items` changes being worked
   out in https://github.com/stac-utils/pystac-client/issues/237. That
   can be done before / after this PR once things are ironed out
2. I changed the `get_all_items` method to use Sphinx's builtin
   `deprecated` directive.
3. I updated the tests (a bit)

I'd like to enable pytest's `-W error` flag to fail on warnings. But
it looks like the `cli` is using the (deprecated) `get_all_items_as_dict`
method. I haven't touched that in this PR, but I don't think the CLI
should be using a deprecated function.