Un-deprecate get_all_items_as_dict
Followup to https://github.com/stac-utils/pystac-client/pull/277 and
closes https://github.com/stac-utils/pystac-client/issues/279

the `cli.search` and `get_all_items()` / `item_collection()` were using
this deprecated method. Users using non-deprecated APIs shouldn't be
getting deprecation warnings.

Clearly it's useful, so let's either un-deprecate it or move the
implementaiton to a private method and call that directly. I've chosen
to un-deprecate it.

I've also configured pytest to fail on warnings, to catch this type of
issue in CI.